+++ changes.md
@@ -1,5 +1,9 @@
+# 2018-08-15
+ + Added private api TRADES and CLIENT_ORDERS requests
+ + Request rate limit is now separated from REST api and increased
+
 # 2018-08-09
- + Added private api BALANCE and BALANCES request
+ + Added private api BALANCE and BALANCES requests
  + Added private api LAST_TRADES request
  * Refreshed python example
  * "Breaking change": Changed type of `sign` field of `WsRequestMetaData` from `string` to `bytes` (you can still use `deprecatedSign`, which is string)